move mongodb classes into separate package
