+++ server/app/db/mongodb/__init__.py
@@ -8,7 +8,7 @@ import json
 import pymongo
 from bson import json_util
 
-from app.db import AlbumMethods
+from app.db import AlbumMethods, ArtistMethods, PlaylistMethods, TrackMethods
 
 
 class Mongo:
@@ -27,6 +27,22 @@ class MongoAlbums(Mongo, AlbumMethods):
         self.collection = self.db["ALL_ALBUMS"]
 
 
+class MongoArtists(Mongo, ArtistMethods):
+    def __init__(self):
+        super(MongoArtists, self).__init__("ALICE_ARTISTS")
+        self.collection = self.db["ALL_ARTISTS"]
+
+
+class MongoPlaylists(Mongo, PlaylistMethods):
+    def __init__(self):
+        super(MongoPlaylists, self).__init__("ALICE_PLAYLISTS")
+        self.collection = self.db["ALL_PLAYLISTS"]
+
+
+class MongoTracks(Mongo, TrackMethods):
+    def __init__(self):
+        super(MongoTracks, self).__init__("ALICE_MUSIC_TRACKS")
+        self.collection = self.db["ALL_TRACKS"]
 # ====================================================================== #
 # cursor convertion methods
