+++ server/app/db/mongodb/albums.py
@@ -2,13 +2,10 @@
 This file contains the Album class for interacting with
 album documents in MongoDB.
 """
-from app.db.mongodb import MongoAlbums
+from app.db.mongodb import MongoAlbums, convert_many, convert_one
 from app import db
 from app.models import Album
 from bson import ObjectId
-
-convert_many = db.convert_many
-convert_one = db.convert_one
 
 
 class Albums(MongoAlbums):
